✨ feat	: followed 상태 추가
